pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mongodb-nl-query-ai-agent"
version = "1.0.0"
description = "AI-powered agent for querying MongoDB databases using natural language"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MongoDB Agent Contributors", email = "praveenkumar.kalamani@gmail.com"}
]
keywords = ["mongodb", "ai", "agent", "natural-language", "llm", "langchain", "database", "query"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pymongo>=4.0.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.1",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["langchain-openai>=0.0.1"]
anthropic = ["langchain-anthropic>=0.0.1"]
bedrock = ["langchain-aws>=0.0.1", "boto3>=1.28.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/praveenk20/mongodb-agent-ai"
Documentation = "https://github.com/praveenk20/mongodb-agent-ai/tree/main/docs"
Repository = "https://github.com/praveenk20/mongodb-agent-ai"
"Bug Tracker" = "https://github.com/praveenk20/mongodb-agent-ai/issues"

[project.scripts]
mongodb-agent = "mongodb_agent.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["mongodb_agent"]
include-package-data = true

[tool.setuptools.package-data]
mongodb_agent = ["semantic_models/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=mongodb_agent --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests (may require external services)",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
